Added casts to get rid of C++ under FreeBSD warnings.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@1741 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/fdcallback.c
@@ -52,7 +52,7 @@ void fdCallbackNew(struct FDCallback* fdCallback,
                    void*              userData)
 {
    struct SimpleRedBlackTreeNode* result;
-   CHECK((fd >= 0) && (fd < FD_SETSIZE));
+   CHECK((fd >= 0) && (fd < (int)FD_SETSIZE));
 
    simpleRedBlackTreeNodeNew(&fdCallback->Node);
    fdCallback->Master          = dispatcher;
